+++ offchain-modules/packages/app-rpc-server/src/client.ts
@@ -79,8 +79,8 @@ export class ForceBridgeAPIV1Client implements ForceBridgeAPIV1 {
         }
         break;
       default:
-        //TODO add other chains
-        void Promise.reject(new Error('not yet'));
+        // TODO add other chains
+        throw new Error('only support Ethereum for now');
     }
     return result;
   }