+++ offchain-modules/packages/x/src/core.ts
@@ -2,6 +2,7 @@ import fs from 'fs';
 import { utils } from '@ckb-lumos/base';
 import { KeyStore } from '@force-bridge/keystore';
 import CKB from '@nervosnetwork/ckb-sdk-core';
+import { ethers } from 'ethers';
 import nconf from 'nconf';
 import { CkbIndexer } from './ckb/tx-helper/indexer';
 import { initLumosConfig } from './ckb/tx-helper/init_lumos_config';
@@ -106,6 +107,8 @@ export class ForceBridgeCore {
   }
 
   async init(config: Config): Promise<ForceBridgeCore> {
+    checkConfigEthereumAddress(config);
+
     // init log
     initLog(config.common.log);
 
@@ -162,4 +165,19 @@ export class ForceBridgeCore {
     ForceBridgeCore._xChainHandler = new XChainHandlers();
     return this;
   }
+}
+
+function checkConfigEthereumAddress(config: Config) {
+  const verifyChecksumAddress = (addresses: string[]) => {
+    addresses.forEach((address) => {
+      if (ethers.utils.getAddress(address) !== address)
+        throw new Error(`${address} is not a ethereum checksum address`);
+    });
+  };
+
+  if (config.eth.contractAddress) verifyChecksumAddress([config.eth.contractAddress]);
+  if (config.eth.multiSignAddresses) verifyChecksumAddress(config.eth.multiSignAddresses);
+  if (config.eth.multiSignHosts) verifyChecksumAddress(config.eth.multiSignHosts.map((value) => value.address));
+  if (config.eth.assetWhiteList) verifyChecksumAddress(config.eth.assetWhiteList.map((value) => value.address));
+  if (config.monitor?.feeAccounts?.ethAddr) verifyChecksumAddress([config.monitor.feeAccounts.ethAddr]);
 }